refactor coach and student ui by creating another component and putting all repeated code in one place
