+++ frontend/src/components/CoachUI.tsx
@@ -1,9 +1,7 @@
 import { useContext, useEffect, useState } from 'react';
 import { observer } from 'mobx-react-lite';
 import { StoreContext } from '../stores/StoreContext';
-import dayjs from 'dayjs';
-import { Card } from 'antd';
-import { Button, Rating, Tooltip, Modal } from '@mantine/core';
+import { Button } from '@mantine/core';
 import { useDisclosure } from '@mantine/hooks';
 import AddTimeSlotModal from './AddTimeSlotModal';
 import '@mantine/dates/styles.css';
@@ -17,13 +15,7 @@ import UserSelector from './UserSelector';
 import Calendar from './Calendar';
 import MeetingFilterModal from './MeetingFilterModal';
 import ReviewModal from './ReviewModal';
-import { formatPhoneNumber } from '../services/formatPhoneNumber';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import {
-  faCalendarDays,
-  faClock,
-  faNoteSticky,
-} from '@fortawesome/free-solid-svg-icons';
+import MeetingCard from './MeetingCard';
 
 const CoachUI = observer(() => {
   const [
@@ -32,9 +24,6 @@ const CoachUI = observer(() => {
   ] = useDisclosure(false);
   const [isReviewOpen, { open: openReview, close: closeReview }] =
     useDisclosure(false);
-  const [isNotesOpen, { open: openNotes, close: closeNotes }] =
-    useDisclosure(false);
-  const [seeNotes, setSeeNotes] = useState(false);
 
   const { coachStore, userStore, timeSlotStore } = useContext(StoreContext);
 
@@ -67,7 +56,6 @@ const CoachUI = observer(() => {
       console.error(`Error creating coach`, error);
     }
   };
-
   const handleReviewClick = (meeting) => {
     openReview();
     timeSlotStore.setTimeSlotUnderReview(meeting);
@@ -97,118 +85,11 @@ const CoachUI = observer(() => {
                 ) : (
                   <div className=' flex flex-col gap-4 mt-4 w-full max-h-[470px] overflow-y-auto'>
                     {coachStore.displayedMeetings.map((meeting) => (
-                      <Card
-                        key={meeting.tsc_id}
-                        // title='Meeting Info'
-                        className={`flex flex-col w-80 ${
-                          meeting.status === 'available'
-                            ? 'border-2 border-blue-600'
-                            : meeting.status === 'booked'
-                            ? 'border-2 border-fuchsia-700'
-                            : ''
-                        } `}
-                      >
-                        <div className='flex gap-4'>
-                          <FontAwesomeIcon
-                            icon={faCalendarDays}
-                            className='self-center'
-                          />
-                          <p className='font-bold'>
-                            {/* <strong>Date:</strong>{' '} */}
-                            {dayjs(meeting.start_time).format(
-                              'dddd, MMMM D, YYYY'
-                            )}
-                          </p>
-                        </div>
-                        <div>
-                          <div className='flex gap-16'>
-                            <p>
-                              <FontAwesomeIcon icon={faClock} />
-                              {dayjs(meeting.start_time).format('HH:mm')}
-                            </p>
-                            <p>
-                              <FontAwesomeIcon icon={faClock} />
-                              {dayjs(meeting.end_time).format('HH:mm')}
-                            </p>
-                          </div>
-
-                          {timeSlotStore.meetingStatus === 'completed' ? (
-                            <div className='flex flex-col'>
-                              <div className='flex justify-between'>
-                                <p className='text-sky-700 font-bold'>
-                                  {meeting.student_name
-                                    ? meeting.student_name
-                                    : ''}
-                                </p>
-                                <Rating
-                                  defaultValue={
-                                    meeting.rating ? meeting.rating : 0
-                                  }
-                                  readOnly
-                                />
-                                <div className='transform transition-transform duration-200 hover:scale-125 cursor-pointer'>
-                                  <Tooltip label='See notes'>
-                                    <FontAwesomeIcon
-                                      icon={faNoteSticky}
-                                      className='text-blue-500'
-                                      onClick={() =>
-                                        coachStore.setNotesDisplay(meeting)
-                                      }
-                                    />
-                                  </Tooltip>
-                                </div>
-                              </div>
-
-                              {coachStore.notesDisplay &&
-                                coachStore.notesDisplay.tsc_id ===
-                                  meeting.tsc_id && (
-                                  <div className='my-2 p-2 border rounded shadow'>
-                                    <p>
-                                      {coachStore.notesDisplay.notes
-                                        ? coachStore.notesDisplay.notes
-                                        : 'No Notes Available'}
-                                    </p>
-                                  </div>
-                                )}
-
-                              <Button
-                                onClick={() => handleReviewClick(meeting)}
-                                size='xs'
-                                variant='light'
-                                className='item-center mt-2'
-                              >
-                                Review
-                              </Button>
-                            </div>
-                          ) : timeSlotStore.meetingStatus === 'booked' ? (
-                            <div className='flex justify-between'>
-                              <div>
-                                {meeting.student_name
-                                  ? meeting.student_name
-                                  : ''}
-                              </div>
-                              <div>
-                                {meeting.student_phone
-                                  ? formatPhoneNumber(meeting.student_phone)
-                                  : ''}
-                              </div>
-                            </div>
-                          ) : (
-                            <div
-                              className={` w-fit px-1 py-0.5 leading-none rounded-md mt-2 ${
-                                meeting.status === 'available'
-                                  ? 'bg-green-200 shadow-md text-green-700'
-                                  : ''
-                              }`}
-                            >
-                              {meeting.status
-                                ? meeting.status
-                                : meeting.dataValues.status}
-                            </div>
-                          )}
-                        </div>
-                      </Card>
-                    ))}{' '}
+                      <MeetingCard
+                        meeting={meeting}
+                        handleReview={handleReviewClick}
+                      />
+                    ))}
                   </div>
                 )}
               </div>
