+++ frontend/src/components/UserSelector.tsx
@@ -1,4 +1,4 @@
-import { useState, useContext } from 'react';
+import { useContext } from 'react';
 import {
   InputBase,
   Combobox,
